[mod/#73] jobId -> Long에서 Int로 변경

+++ app/src/main/java/com/haphap/app/presentation/jobdetail/navigation/JobDetailNavigation.kt
@@ -27,7 +27,7 @@ fun NavGraphBuilder.jobDetailGraph(
         JobDetailRoute(
             navigateBack = { navController.popBackStack() },
             navigateToRegister = {
-                navController.navigateToRegisterFromJobDetail(jobId = route.postingId.toLong())
+                navController.navigateToRegisterFromJobDetail(jobId = route.postingId)
             },
             modifier = Modifier.padding(innerPadding),
         )
